+++ qt/base_qt.py
@@ -1,5 +1,4 @@
 from functools import partial
-
 from PyQt5.QtWidgets import QTabWidget, QWidget
 
 
@@ -18,6 +17,9 @@ def wrap(cls):
 @wrap
 class BaseQt:
     """通过.语法调用子控件，把所有控件都命名会很乱"""
+
+    def __init__(self): ...
+
     def add_tab(self: QWidget, p1, p2, name=''):
         setattr(self, name, p1)
         QTabWidget.addTab(self, p1, p2)
